+++ project/backend-services/smart-irrigation-backend/domain-services/src/main/java/pt/sensae/services/smart/irrigation/backend/domainservices/garden/model/BoundaryCommandDetails.java
@@ -4,9 +4,9 @@ public class BoundaryCommandDetails {
 
     public int position;
 
-    public Float latitude;
+    public Double latitude;
 
-    public Float longitude;
+    public Double longitude;
 
-    public Float altitude;
+    public Double altitude;
 }
