+++ project/backend-services/smart-irrigation-backend/domain/src/main/java/pt/sensae/services/smart/irrigation/backend/domain/model/GPSPoint.java
@@ -2,17 +2,17 @@ package pt.sensae.services.smart.irrigation.backend.domain.model;
 
 import pt.sharespot.iot.core.sensor.data.GPSDataDTO;
 
-public record GPSPoint(Float latitude, Float longitude, Float altitude) {
+public record GPSPoint(Double latitude, Double longitude, Double altitude) {
 
     public static GPSPoint from(GPSDataDTO data) {
-        return new GPSPoint(data.latitude.floatValue(), data.longitude.floatValue(), data.altitude.floatValue());
+        return new GPSPoint(data.latitude, data.longitude, data.altitude);
     }
 
-    public static GPSPoint ofLatLong(Float latitude, Float longitude) {
+    public static GPSPoint ofLatLong(Double latitude, Double longitude) {
         return new GPSPoint(latitude, longitude, null);
     }
 
-    public static GPSPoint ofLatLongAlt(Float latitude, Float longitude, Float altitude) {
+    public static GPSPoint ofLatLongAlt(Double latitude, Double longitude, Double altitude) {
         return new GPSPoint(latitude, longitude, altitude);
     }
 }
