+++ project/backend-services/smart-irrigation-backend/infrastructure/persistence/postgres/src/main/java/pt/sensae/services/smart/irrigation/backend/infrastructure/persistence/postgres/mapper/gardeningArea/AreaBoundariesMapper.java
@@ -27,9 +27,9 @@ public class AreaBoundariesMapper {
 
     public static Area daoToModel(Stream<AreaBoundariesPostgres> points) {
         var collect = points.map(dao -> {
-            var aFloat = dao.altitude != null ? Float.parseFloat(dao.altitude) : null;
+            var aFloat = dao.altitude != null ? Double.parseDouble(dao.altitude) : null;
 
-            var gpsPoint = new GPSPoint(Float.parseFloat(dao.latitude), Float.parseFloat(dao.longitude), aFloat);
+            var gpsPoint = new GPSPoint(Double.parseDouble(dao.latitude), Double.parseDouble(dao.longitude), aFloat);
             return new BoundaryPoint(dao.position, gpsPoint);
         }).collect(Collectors.toSet());
         return new Area(collect);
